refactor: extract telegram request intake
